Working on JS/Storage integration
- Fix decoders/encoders for BC
- Working ports for settings
