+++ src/js/logs.js
@@ -1,8 +1,4 @@
 import db from "./helpers/db.js";
-import download from "downloadjs";
-import { peakImportFile } from "dexie-export-import";
-
-import setFlash from "./helpers/flash.js";
 
 const insert = cycle => db.cycles.add(cycle);
 
@@ -17,29 +13,11 @@ const monthlyLogs = millis => {
 const fetch = async (app, millis) => {
   const logs = await monthlyLogs(millis);
 
-  app.ports.gotStatsLogs.send({ ts: millis, logs: logs });
-};
-
-const exportData = async () => {
-  const blob = await db.export();
-
-  download(blob, "pelmodoro-data.json", "application/json")
+  app.ports.gotFromLog.send({ ts: millis, logs: logs });
 };
 
 
-const importData = async (app, str) => {
-  const blob = new Blob([str]);
 
-  try {
-    await peakImportFile(blob);
-    await db.cycles.clear();
-    await db.import(blob);
-
-    setFlash(app, "Success", "Data has been successfully imported.")
-  } catch (e) {
-    setFlash(app, "Error", "There was an error trying to import the data.")
-  }
-}
 
 const updateCycle = async data => {
   const cycle = await db.cycles.where({ start: data[0]}).toArray();
@@ -53,21 +31,21 @@ const updateCycle = async data => {
   return updated;
 }
 
-const clearLogs = () => {
-  const confirmed = confirm("Are you sure you wanna delete all log entries? This step is irreversible!");
-
-  if (!confirmed) {
-    return;
-  }
-
-  db.cycles.clear();
-}
 
 export default function (app) {
-  app.ports.logCycle.subscribe(insert);
-  app.ports.fetchLogs.subscribe(async millis => await fetch(app, millis));
-  app.ports.requestDataExport.subscribe(async () => await exportData());
-  app.ports.importData.subscribe(async blob => await importData(app, blob));
-  app.ports.updateCycle.subscribe(async data => await updateCycle(data));
-  app.ports.clearLogs.subscribe(clearLogs);
+  app.ports.toLog.subscribe(async data => {
+    switch (data["type"]) {
+      case "fetch":
+        await fetch(app, data["time"]);
+        break;
+    }
+  });
+
+
+  // app.ports.logCycle.subscribe(insert);
+  // app.ports.fetchLogs.subscribe(async millis => await fetch(app, millis));
+  // app.ports.requestDataExport.subscribe(async () => await exportData());
+  // app.ports.importData.subscribe(async blob => await importData(app, blob));
+  // app.ports.updateCycle.subscribe(async data => await updateCycle(data));
+  // app.ports.clearLogs.subscribe(clearLogs);
 }
